requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile app/pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via asyncpraw
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   asyncpraw
    #   asyncprawcore
    #   langchain-community
    #   py-cord
    #   tweepy
    #   twitchio
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.17.0
    # via asyncpraw
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   groq
    #   httpx
async-lru==1.0.3
    # via tweepy
asyncpraw==7.8.1
    # via pydiscogs
asyncprawcore==2.4.0
    # via asyncpraw
attrs==25.4.0
    # via aiohttp
audioop-lts==0.2.2
    # via pydiscogs
beautifulsoup4==4.14.2
    # via yfinance
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   polygon-api-client
    #   requests
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via duckduckgo-search
curl-cffi==0.13.0
    # via yfinance
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via groq
duckduckgo-search==7.5.5
    # via pydiscogs
filetype==1.2.0
    # via langchain-google-genai
frozendict==2.4.6
    # via yfinance
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.186.0
    # via google-generativeai
google-auth==2.42.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-genai==1.20.0
    # via pydiscogs
google-generativeai==0.8.5
    # via langchain-google-genai
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
groq==0.33.0
    # via langchain-groq
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   groq
    #   langgraph-sdk
    #   langsmith
    #   ollama
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via langchain-community
langchain-community==0.3.20
    # via pydiscogs
langchain-core==0.3.79
    # via
    #   langchain
    #   langchain-community
    #   langchain-google-genai
    #   langchain-groq
    #   langchain-ollama
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==2.0.10
    # via pydiscogs
langchain-groq==0.2.3
    # via pydiscogs
langchain-ollama==0.2.2
    # via pydiscogs
langchain-text-splitters==0.3.11
    # via langchain
langgraph==0.3.21
    # via pydiscogs
langgraph-checkpoint==2.1.2
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.1.8
    # via langgraph
langgraph-sdk==0.1.74
    # via langgraph
langsmith==0.3.45
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==6.0.2
    # via duckduckgo-search
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.12
    # via yfinance
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.4
    # via
    #   langchain-community
    #   pandas
    #   yfinance
oauthlib==3.3.1
    # via
    #   requests-oauthlib
    #   tweepy
ollama==0.5.1
    # via
    #   langchain-ollama
    #   pydiscogs
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pandas==2.3.3
    # via yfinance
peewee==3.18.2
    # via yfinance
platformdirs==4.5.0
    # via yfinance
polygon-api-client==1.14.6
    # via pydiscogs
primp==0.15.0
    # via duckduckgo-search
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   yfinance
py-cord==2.6.1
    # via
    #   app (app/pyproject.toml)
    #   pydiscogs
pyaml-env==1.2.1
    # via pydiscogs
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   google-genai
    #   google-generativeai
    #   groq
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   ollama
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via langchain-community
pydiscogs @ git+https://github.com/bpafoshizle/pydiscogs@41ba47fd6ac54942c4555027c33e19f768bf8fde
    # via app (app/pyproject.toml)
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   app (app/pyproject.toml)
    #   pydantic-settings
    #   pydiscogs
pytz==2025.2
    # via
    #   pandas
    #   pydiscogs
    #   yfinance
pyyaml==6.0.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   pyaml-env
requests==2.32.5
    # via
    #   google-api-core
    #   google-genai
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-oauthlib
    #   requests-toolbelt
    #   tweepy
    #   update-checker
    #   yfinance
requests-oauthlib==1.3.1
    # via tweepy
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   groq
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via
    #   langchain
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tqdm==4.67.1
    # via google-generativeai
tweepy==4.12.1
    # via pydiscogs
twitchio==3.1.0
    # via pydiscogs
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   beautifulsoup4
    #   google-genai
    #   google-generativeai
    #   groq
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
update-checker==0.18.0
    # via asyncpraw
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   polygon-api-client
    #   requests
uvloop==0.22.1
    # via pydiscogs
websockets==14.2
    # via
    #   google-genai
    #   polygon-api-client
    #   yfinance
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via
    #   aiohttp
    #   asyncprawcore
yfinance==0.2.63
    # via pydiscogs
zstandard==0.23.0
    # via langsmith
